Added configuration "menu"
Created the Configurer, a class that allows settings to be changed/output to be converted and compressed, all through the use of the two buttons and the feedback of the LEDs. Still a WIP, but off to a great start.
Also added connector.py, the start of a class that deals with internet connections and sending data, as well as a pi-shutdown script used by both the Configurer and Controller.